Made changes to fix bug where UI would be weird.
Made changes to fix bug where if you went from youtube's front page or a
search page to a video, the UI would not load until you did a refresh.
Unknown if this bug was always there or was introduced at some point; I
always just went to a video page for testing.