Removed redundant video player object.
There was a video player object (in youtubejt_cs) that took a plain
video object and encapsulated certain functions. I've decided to cut
that out and replace it with calling to the video object directly with
parameters gleaned from data attributes retrieved on the front end. This
removes redundancy when adding new features to the base video player and
opens the door for extra options. All changes in this commit radiate
from that.